Update yt-dlp.exe to version 2025.05.22
Updated the version of the `yt-dlp.exe` executable in the `.gitignore`, `README.md`, and `ytdlp.executable.csproj` files from `v2025.04.30` to `v2025.05.22`.

In the `ytdlp.executable.csproj`, added a `<CopyToOutputDirectory>` element set to `Always` for the `yt-dlp.exe` file, ensuring it is always copied to the output directory during the build process.